refactor : (main) 레이아웃으로 헤더, 패딩 통합 - (main)/layout.tsx 생성   - 모든 main 페이지에 공통 적용 : Header + Padding - home, history, ranking, interview, setting을 (main) route group으로 이동
