+++ src/app/(main)/history/page.tsx
@@ -1,0 +1,9 @@
+import HistorySection from '@/components/history/HistorySection'
+
+export default function HistoryPage() {
+  return (
+    <div className="flex-1 min-h-0 flex flex-col overflow-hidden pb-10">
+      <HistorySection />
+    </div>
+  )
+}
