+++ src/app/(main)/interview/job-posting/[uuid]/analysis/page.tsx
@@ -1,0 +1,5 @@
+import AnalysisSection from '@/components/interview/AnalysisSection'
+
+export default function AnalysisPage() {
+  return <AnalysisSection />
+}
