+++ src/app/(main)/interview/job-posting/[uuid]/complete/page.tsx
@@ -1,4 +1,3 @@
-import Header from '@/components/common/header/Header'
 import CompleteSection from '@/components/interview/CompleteSection'
 import { TOTAL_QUESTIONS } from '@/utils/interview'
 
@@ -11,12 +10,9 @@ export default async function CompletePage({
   const currentQuestion = Math.max(1, parseInt(q ?? '1', 10) || 1)
 
   return (
-    <main className="h-screen flex flex-col overflow-hidden">
-      <Header />
-      <CompleteSection
-        currentQuestion={currentQuestion}
-        totalQuestions={TOTAL_QUESTIONS}
-      />
-    </main>
+    <CompleteSection
+      currentQuestion={currentQuestion}
+      totalQuestions={TOTAL_QUESTIONS}
+    />
   )
 }
