+++ src/app/(main)/interview/job-posting/[uuid]/question/page.tsx
@@ -1,5 +1,4 @@
 import { notFound } from 'next/navigation'
-import Header from '@/components/common/header/Header'
 import QuestionSection from '@/components/interview/QuestionSection'
 import { parseQuestionNumber } from '@/utils/interview'
 
@@ -12,10 +11,5 @@ export default async function QuestionPage({
   const questionNumber = parseQuestionNumber(q)
   if (questionNumber === null) notFound()
 
-  return (
-    <main className="h-screen flex flex-col overflow-hidden">
-      <Header />
-      <QuestionSection questionNumber={questionNumber} />
-    </main>
-  )
+  return <QuestionSection questionNumber={questionNumber} />
 }
