+++ src/app/(main)/interview/job-posting/page.tsx
@@ -1,0 +1,5 @@
+import JobPostingFormSection from '@/components/interview/JobPostingFormSection'
+
+export default function JobPostingPage() {
+  return <JobPostingFormSection />
+}
