+++ src/app/(main)/interview/page.tsx
@@ -1,0 +1,5 @@
+import InterviewStartSection from '@/components/interview/InterviewStartSection'
+
+export default function InterviewPage() {
+  return <InterviewStartSection />
+}
