+++ src/app/(main)/ranking/page.tsx
@@ -1,0 +1,5 @@
+import RankingSection from '@/components/ranking/RankingSection'
+
+export default function RankingPage() {
+  return <RankingSection />
+}
