--- src/app/interview/job-posting/[uuid]/analysis/page.tsx
@@ -1,11 +1,0 @@
-import Header from '@/components/common/header/Header'
-import AnalysisSection from '@/components/interview/AnalysisSection'
-
-export default function AnalysisPage() {
-  return (
-    <main className="h-screen flex flex-col overflow-hidden">
-      <Header />
-      <AnalysisSection />
-    </main>
-  )
-}
