+++ src/components/interview/JobPostingFormSection.tsx
@@ -79,7 +79,7 @@ export default function JobPostingFormSection() {
   const isComplete = urlValid
 
   return (
-    <section className="flex flex-col flex-1 min-h-0 gap-5 px-10 pt-6.5 pb-10">
+    <section className="flex flex-col flex-1 min-h-0 gap-5 pt-6.5 pb-10">
       <h1 className="h1 w-full">공고 맞춤 면접 연습</h1>
 
       <div className="flex-1 min-h-0 flex flex-col items-center justify-center w-full min-w-200 bg-secondary rounded-2xl border border-primary">
