+++ src/components/ranking/RankingSection.tsx
@@ -12,8 +12,8 @@ export default function RankingSection() {
   })
 
   return (
-    <div className="flex-1 min-h-0 overflow-auto w-full max-w-360 mx-auto px-[clamp(16px,2.5vw,80px)]">
-      <section className="p-8">
+    <div className="flex-1 min-h-0 overflow-auto">
+      <section className="py-8">
         <div className="h1 mb-4">랭킹을 확인해보세요</div>
         <div className="p2 mb-2">나의 순위</div>
         {data?.myRankingResponse && (
@@ -24,7 +24,7 @@ export default function RankingSection() {
           />
         )}
       </section>
-      <div className="flex flex-col px-10">
+      <div className="flex flex-col">
         <div className="flex items-center gap-2 py-4">
           <span className="p2 text-gray-1">전체 유저 순위</span>
           <span className="p4 text-gray-3">
